An additional series of options is presented to the user so that an already accurate value of the time period can be used to get a single phase folded plot for a system.
After phase folding by iterating the period the user is now given the option to save the accurate time period to a csv file (System_Parameters.csv) which stores catalogue number and accurate period in columns 1 and 2 respectively.
During a save the code looks to see if a value exists already and will overwrite the previous period estimate.

In a subsequent build I hope to change the time period input routine to first look up a value in the csv file since if an accurate time period exists already then it should be there.  But it will fall back to asking the user for the value.

An additional function to transpose 2D rows into columns was added and previous transpose function name was amended to make clear that it transposes columns into rows.